Add flow control module
