+++ mypy.ini
@@ -291,9 +291,6 @@ disallow_untyped_calls = False
 disallow_untyped_defs = False
 disallow_untyped_calls = False
 
-[mypy-wandb.sdk.internal.writer]
-ignore_errors = True
-
 [mypy-wandb.sdk.internal.datastore]
 ignore_errors = True
