+++ wandb/sdk/internal/settings_static.py
@@ -50,6 +50,7 @@ class SettingsStatic:
     anonymous: Optional[str]
     host: Optional[str]
     username: Optional[str]
+    sync_file: str
 
     # TODO(jhr): clean this up, it is only in SettingsStatic and not in Settings
     _log_level: int
